# ByteMarket/backend/src/main/java/es/grupo04/backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/v1/profile:
    put:
      summary: Update user profile information and optionally upload a profile picture
      tags:
      - profile-rest-controller
      operationId: updateProfile
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: address
        in: query
        required: true
        schema:
          type: string
      - name: newPass
        in: query
        required: true
        schema:
          type: string
      - name: repeatPass
        in: query
        required: true
        schema:
          type: string
      - name: iframe
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                profilePicInput:
                  type: string
                  format: binary
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/users/{userId}/reviews:
    get:
      summary: Fetch all reviews for a specific user by their ID
      tags:
      - review-report-rest-controller
      operationId: getReviews
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewDTO"
    post:
      summary: Submit a new review for a user by their ID
      tags:
      - review-report-rest-controller
      operationId: postReview
      parameters:
      - name: reviewDTO
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NewReviewDTO"
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
  /api/v1/users/signin:
    post:
      summary: Register a new user
      tags:
      - user-rest-controller
      operationId: createUser
      parameters:
      - name: user
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NewUserDTO"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/reports/products/{productId}:
    post:
      summary: Report a product by its ID
      tags:
      - review-report-rest-controller
      operationId: postReport
      parameters:
      - name: reportDTO
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NewReportDTO"
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReportDTO"
...
  /api/v1/users:
    get:
      summary: Retrieve a list of all users
      tags:
      - user-rest-controller
      operationId: getAllUsers
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
  /api/v1/users/{userId}:
    get:
      summary: Fetch a user’s details by their ID
      tags:
      - user-rest-controller
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/users/me:
    get:
      summary: Retrieve details of the authenticated user
      tags:
      - user-rest-controller
      operationId: getAuthenticatedUser
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/reviews/{reviewId}:
    get:
      summary: Retrieve a review by its ID
      tags:
      - review-report-rest-controller
      operationId: getReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
    delete:
      summary: Delete a review by its ID
      tags:
      - review-report-rest-controller
      operationId: deleteReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/reports:
    get:
      summary: Fetch a list of all reports
      tags:
      - review-report-rest-controller
      operationId: getReports
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportDTO"
  /api/v1/profile/stats:
    get:
      summary: Retrieve statistics for the user’s profile
      tags:
      - profile-rest-controller
      operationId: stats
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/profile/image/{id}:
    get:
      summary: Fetch the profile image of a user by their ID
      tags:
      - profile-rest-controller
      operationId: getProfileImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/reports/{reportId}:
    delete:
      summary: Delete a report by its ID
      tags:
      - review-report-rest-controller
      operationId: deleteReport
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/profile/{id}:
    delete:
      summary: Delete a user account by its ID
      tags:
      - profile-rest-controller
      operationId: deleteAccount
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        image:
          type: string
        hasImage:
          type: boolean
    NewReviewDTO:
      type: object
      properties:
        rating:
          type: integer
          format: int32
        description:
          type: string
        reviewedUserId:
          type: integer
          format: int64
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        reviewOwner:
          $ref: "#/components/schemas/UserBasicDTO"
        reviewedUser:
          $ref: "#/components/schemas/UserBasicDTO"
        rating:
          type: integer
          format: int32
    NewUserDTO:
      type: object
      properties:
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        repeatPassword:
          type: string
    NewReportDTO:
      type: object
      properties:
        reason:
          type: string
        description:
          type: string
        productId:
          type: integer
          format: int64
        reportCreatorId:
          type: integer
          format: int64
    ProductBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        sold:
          type: boolean
        publishDate:
          type: string
          format: date
    ReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        reason:
          type: string
        description:
          type: string
        product:
          $ref: "#/components/schemas/ProductBasicDTO"
        reportCreator:
          $ref: "#/components/schemas/UserBasicDTO"
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        creationYear:
          type: integer
          format: int32
        roles:
          type: array
          items:
            type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductBasicDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        sales:
          type: array
          items:
            $ref: "#/components/schemas/ProductBasicDTO"
        purchases:
          type: array
          items:
            $ref: "#/components/schemas/ProductBasicDTO"
        iframe:
          type: string
        image:
          type: string
        hasImage:
          type: boolean